Allow Nginx auth proxy to optionally be served on non-standard ports
In order to make the developer experience nicer, we think that it should
be possible to run the Nginx auth proxy on an unprivileged port (i.e.
>1024) - as running the container on the standard port 80/443 requires
the docker daemon to be running with root privileges.

This change allows port numbers to be specified with the
`$CJSE_NGINX_PORT_HTTP` and `$CJSE_NGINX_PORT_HTTPS` environment
variables, which default to 80 and 443 if not specified, respectively.